feat: improve chart and insight display in reports
Now, I'd like to improve the data display in the charts tab. Each report_item in Nightscout.tsx is actually of type ReportItem, which is made up of a chart and a set of AssessmentInsights. I'd like to create a component that encapsulates displaying the chart and then all of the insights underneath. Don't write code right away, let's walk through this as if we were two product managers and a UX designer coming up with product requirements and look/feel.

For each rendering of the component, I'd like to show the chart on top, and the insights underneath. The insights should be displayed in the stable order they are written in. However, they may be of different InsightPriority enum values.

I'd like CRITICAL and SERIOUS insights to somehow be differentiated. I don't want it to be too much, but I was thinking maybe prefixing a light bulb icon for ALWAYS_INCLUDE, with a red error icon for CRITICAL, a yellow error icon for SERIOUS, a grey/black warning icon for IMPORTANT, and a bullet point for everything else.

A chart can have between 0-5 insights. More than that would be exceptional.

Does this make sense? What else should I consider?

```git-revs
f001ba6  (Base revision)
aa8949c  Create the ReportItemDisplay component with chart and insights
8572d98  Snapshot before codemcp change
d9e03d8  Update Nightscout component to use ReportItemDisplay
HEAD     Add import for ReportItemDisplay component
```

codemcp-id: 15-feat-improve-chart-and-insight-display-in-reports