feat: add optional mealevents parameter to ClusterEventsChart
I'd like ClusterEventsChart to take an optional parameter for mealevents. I'd like to pass in a variable called treatments of type NightscoutTreatment[] (which is already defined in @/types/nightscout)

To plot these, it should be an optional bar chart subplot underneath the current timeseries for entries (which is the blood glucose). Ideally the meals should be colored for the same Event in the time series (ie, essentially the same day, since each event is a different day).

```git-revs
7eea580  (Base revision)
1d5cc2f  Adding NightscoutTreatment import
b9518b1  Adding treatments parameter to ClusterEventsChartProps interface
1744a25  Adding treatments parameter to function signature
c0780c0  Adding treatments to useEffect dependency array
HEAD     Adding helper functions for processing meal data
```

codemcp-id: 51-feat-add-optional-mealevents-parameter-to-clustere