feat: review and plan for static Nightscout demo page
I have a next.js 14 webapp that has a page (Nightscout.tsx) that has a web form that, when called, retrieves a bunch of data that is then shown on the screen, including charts and textual data.

I want to take this page and save a statically saved version of it, where all the data is available as a 'demo' page to the user. It shouldn't require the use of any API to call data- the data should all be embedded in the website, or included in adjacent, imported files.

Please review Nightscout.tsx and recommend options for doing this. Please note a lot of the data is currently in localStorage. Part of the challenge is getting a one-time extraction and putting this into a demo page so users can see the reports.

Don't implement anything, just evaluate and report your findings.

```git-revs
117cdc8  (Base revision)
HEAD     Create initial copy of Nightscout.tsx
```

codemcp-id: 62-feat-review-and-plan-for-static-nightscout-demo-pa